Add support to convert single character lexemes to Tokens
** Description **
This commit scans single character, and adds the appropriate
TokenType for it.

Lexeme: A lexeme is a sequence of characters in the source program
that represent something in a program. They are raw strings in the
source code.

Token in an instance of lexeme with more details like literal, type etc.

** Motivation **

Scan single characters like (, ), {, }  etc.,