Generate Expr Class and validate the tree using AstPrinter
** Description **

Generate the Expr.java class using the helper script GenerateAst.
Also, add support for the AST print to check if we are parse the
given Expression.

** Motivation **

Generate parse tree, and will be useful in case we need to check
precendence and associativity of the expression, and if our parser
is following the same.

** Testing **

Ran the manual generate Binary Expression in AstPrinter.main()
Output: (* (- 123) (group 45.67))