Add state for storing the variable declaration values
** Description **

This commit stores the state of the variable defined using a
Hashmap.

Interpreter, when it find the variable declaration, it evaluates it
by storing or defining the value in the hashmap.

When Interpreter finds the variable expresions (var a = b or print b)
Then it gets the value from the hashmap.

Notes: We thow runtime error if we dont get the variable defined
and not static or parse error. This is by design choice.