Add Support for Function Call Parsing and Evaluation
** Description **

This commit adds support to parse the function call, and also provides
implementation details on how to evaluate these function.

Commit also check "arity", the number of arguments in function call
is same as function parameters.

Native Functions, are functions that are exposed by interpreter
that are implemented in the Host language (here Java). One such
function we are providing is the "clock" function to measure time
and benchmarks.

We also add couple of sanity checks like we can only call function
if its a callable.