Add function declaration Parsing and evaluation
** Description **

Note:
  - Function Declaration STATEMENT
  - Function Call EXPRESSION

Add support for declartion function, parameters and also function.
Also, evaluate the function body which is a list of statements.

Also, add support for returning the value from the function. We achieve
this by taking advantage of JVM Stack, that is by throwing a custom Exception
and catching where the call was made, and then returing the value.