Add Variable Resolution Pass
** Description **

This commit adds Variable Resolution pass which is in between
parser and interpreter to determine which variable a specific
declaration, assignment or use refer to.

** Motivation **

During closure, we added a change to capture the surrounding environment,
and these environment or keep changing. Hence, technically the variable
passed during closure during declaration should be the one that get used.

One of the approach is using persistent environment, where effectively we take a
screenshot of the environment, and pass it to closure, and it uses the same env
and it not affected by the changing env.

However, we use a different approach here of checking how many hops the variable
is declared, and we while interpreting, we jump that many hops to get the variable.
That was we always reference or resolve the same variable.

Also, since we are using static scoping, variable need to resolved once,
and we always refer to the same variable.